系统      第一次提交到git，7月16日

+++ crm1/src/main/java/com/sc/mapper/XtUserAccountMapper.java
@@ -2,6 +2,8 @@ package com.sc.mapper;
 
 import com.sc.entity.XtUserAccount;
 import com.sc.entity.XtUserAccountExample;
+
+import java.math.BigDecimal;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
@@ -10,15 +12,19 @@ public interface XtUserAccountMapper {
 
     int deleteByExample(XtUserAccountExample example);
 
-    int deleteByPrimaryKey(Long userId);
+    
+    //删除
+    int deleteByPrimaryKey(BigDecimal userId);
 
+    //添加
     int insert(XtUserAccount record);
 
     int insertSelective(XtUserAccount record);
 
     List<XtUserAccount> selectByExample(XtUserAccountExample example);
 
-    XtUserAccount selectByPrimaryKey(Long userId);
+    //通过id做查询
+    XtUserAccount selectByPrimaryKey(BigDecimal userId);
 
     int updateByExampleSelective(@Param("record") XtUserAccount record, @Param("example") XtUserAccountExample example);
 
@@ -26,5 +32,6 @@ public interface XtUserAccountMapper {
 
     int updateByPrimaryKeySelective(XtUserAccount record);
 
+    //修改
     int updateByPrimaryKey(XtUserAccount record);
 }
